+++ dmr/controller.py
@@ -255,7 +255,7 @@ class Blueprint(Generic[_SerializerT_co]):  # noqa: WPS214
         self,
         error: str | Exception,
         *,
-        loc: str | None = None,
+        loc: str | list[str | int] | None = None,
         error_type: str | ErrorType | None = None,
     ) -> Any:  # `Any`, so we can change the return type in subclasses.
         """
@@ -265,7 +265,8 @@ class Blueprint(Generic[_SerializerT_co]):  # noqa: WPS214
             error: A serialization exception like a validation error or
                 a ``dmr.exceptions.DataParsingError``.
             loc: Location where this error happened.
-                Like "headers" or "field_name".
+                Like ``"headers"``, or ``"field_name"``,
+                or ``["parsed_headers", "header_name"]``.
             error_type: Optional type of the error for extra metadata.
 
         Returns:
@@ -620,6 +621,7 @@ class Controller(Blueprint[_SerializerT_co], View):  # noqa: WPS214
             method.lower(): endpoint.get_schema(
                 path,
                 pattern,
+                cls.__qualname__,
                 cls.serializer,
                 context,
             )