+++ docs/examples/error_handling/custom_error_messages.py
@@ -22,7 +22,7 @@ class _CustomErrorMixin:
         self,
         error: str | Exception,
         *,
-        loc: str | None = None,
+        loc: str | list[str | int] | None = None,
         error_type: str | ErrorType | None = None,
     ) -> CustomErrorModel:
         default = format_error(