+++ docs/examples/routing/blueprints.py
@@ -24,3 +24,6 @@ router = Router(
 urlpatterns = [
     path(router.prefix, include((router.urls, 'server'), namespace='api')),
 ]
+
+# run: {"method": "get", "url": "/api/user/", "use_urlpatterns": true}  # noqa: ERA001, E501
+# run: {"method": "post", "url": "/api/user/", "body": {"email": "some@example.com", "age": 20}, "use_urlpatterns": true}  # noqa: ERA001, E501