+++ tests/test_unit/test_openapi/test_example_snapshots.py
@@ -40,17 +40,7 @@ class _UserController(
         raise NotImplementedError
 
 
-@pytest.mark.freeze_time(
-    dt.datetime(
-        year=2026,  # noqa: WPS432
-        month=3,
-        day=10,
-        hour=5,
-        minute=2,
-        second=5,
-        tzinfo=dt.UTC,
-    ),
-)
+@pytest.mark.freeze_time('02-11-2025 10:15:00')
 def test_user_schema_with_examples(
     snapshot: SnapshotAssertion,
     dmr_clean_settings: None,