+++ tests/test_unit/test_plugins/test_pydantic/test_error_model_customization.py
@@ -47,7 +47,7 @@ class _CustomErrorMixin:
         self,
         error: str | Exception,
         *,
-        loc: str | None = None,
+        loc: str | list[str | int] | None = None,
         error_type: str | ErrorType | None = None,
     ) -> _CustomErrorModel:
         default = format_error(